Created a mercurial repository for the 'Notions' projects group.
During the first 3 months of 2009 I created 6 distinct applications and a library.
At the time, I was unaware of the existence of revision control systems so only
the last version of each project is preserved.

Though each project was designed and implemented independently, they were
created in chronological order and each of them builds on the knowledge and
experience gained from the previous ones. Moreover, they share some common
source.
Thus, now that I am recreating the history, I decided to encode them as
different branches of a single repository.

Each project had its own name which will be used as the name of the branch,
while the name of the last one will be used as an umbrella name for the whole
group.