Added original VS2008 project to hg after renaming it to 'Notions' from 'Mark III'.
This version has the same functionality as the previous one (Mark II) but uses
DirectInput8 instead of glut for acquiring mouse and keyboard input.

This version uses the 'intThrower' hack to bypass glut's main loop and is
unlikely to work on Windows 7 or later operating systems.

Most of the changes are in main.cpp to replace the main loop. Essentially input
is acquired via DirectInput8's polling functions and a suitable handler from the
previous version is called.

Note:
The use of DirectInput limits portability, but at the time it was deemed a
necessary step on the road to support high resolution pen input (and pen
buttons, eraser, etc. none of which glut supported). Of course that goal was
never achieved due to lack of time and hardware failure...

--HG--
branch : Mark III