Added original VS2008 project to hg after renaming it to 'Notions' from 'Mark III thread'.
This version is essentially the same as Mark III but uses Boost::Thread to
process input in one thread while rendering on another. This attempts to solve
the problem of not receiving input while the rendering thread waits for the
vertical synchronization signal (e.g. in glutSwapBuffers).

This version also uses the 'intThrower' hack to bypass glut's main loop and is
unlikely to work on Windows 7 or later operating systems.

Note:
At the time I didn't know how to synchronize access to resources shared between
threads so this version didn't go far...

--HG--
branch : Mark III Thread