Added original VS2008 project to hg after renaming it to 'Notions' from 'display lists'.
This version lets you draw a free-form curve with the mouse and displays, at
real time, the best straight line approximating the points of the curve
calculated using total least squares. A grid and two bent two-way arrows on a
circle are displayed. You can left-click on any one of the arrows and drag it in
a circular manner to rotate the arrows and the grid around their center.
(when you drag the arrows you continue to draw... sorry!)

Note that you can't build this project without an installation of the
'Unsorted Library' and its behavior may change with the version of the library.
Earlier versions of the library used the 'intThrower' hack to bypass glut's main
loop and are unlikely to work on Windows 7 or later operating systems.

Note:
Starting with Mark III I began diverging functionality from the main program to
a library (now the 'Unsorted Library' project). Mark III used some classes, Mark
III thread used some more... by the time 'display lists' was written, it only
needed a main! In fact the project was originally created to test components of
the library such as 'SquareGrid' and 'RotationIndicator' implemented using
OpenGL display lists.

--HG--
branch : Display Lists