Added original VS2008 project to hg after renaming it to 'Notions' from 'notions'.
This is a full-blown version. It lets you draw free curves that will stay as
they are and others that will be converted to polylines and polygons (convex or
not) it lets you select them, modify them and move them around. It allows you to
embed unconverted curves into converted convex polygons. It also allows you to
pan, rotate and center the whole 'drawing surface'. It displays a grid and a
minimalistic user interface that lets you concentrate in your drawing.
The user interaction is based on 'modes' that are controlled with the keyboard
arrows. The modes are 'conversion' (default), 'selection', 'navigation',
'creation' and 'configuration' (not used yet) and can be made "sticky" by
pressing enter. The current mode, zoom and time are displayed on screen. In
navigation mode two bent two-way arrows are displayed on a circle (the ones from
the 'Display Lists' version) together with a indication of the current rotation.
The user can pan the drawing surface by dragging the part inside the arrows,
zoom it in/out by dragging the arrows themselves and rotate it by dragging the
part outside the arrows.
The 'navigation' interface was a success by any measure! :)
One only really appreciates it when he uses it with a digital pen!
I does more like enabling you to connect the polygons with polylines and
disconnect them. I'm probably not remembering it all! :p

This version has essentially no common source base with the other versions.
It uses all (or nearly all) the library functionality created along the way
though!

The same cautions about the library version and the 'intThrower' hack that were
mentioned in 'Display Lists' apply here as well.

--HG--
branch : Notions