Fixed 2 bugs and did some general clean-up.
Fixed a bug causing a crash when the down arrow was pressed.

The crash was caused by dereferencing an iterator beyond the end of the
freePoints vector. This in turn was caused by not clearing velocity_magnitudes,
velocities and line_stats vectors at the right time.


Fixed a bug causing incorrect coloring of points in COLORED_POINTS_CONTINUOUS
mode.

This was caused by using uninitialized min and max global variables.
Still, using a single min, max and threshold value for coloring all strokes but
a different one for each stroke when calculating its segments and polyline is
not correct.


Commented code used for logging velocity magnitudes.

--HG--
branch : Mark II